[XLA:GPU][NFC] Consolidate CuTeDSL FFI implementation
